Added Support for Nesting Vertical in Vertical and Horizontal in Horizontal
Added support to allow for nesting a vertical slider panel within a vertical slider panel. Added a sample utilizing this functionality in the Blazor Server and Blazor WAMS projects. Added an integration test to test it.
Added support to allow for nesting a horizontal slider panel within a horizontal slider panel. Added a sample utilizing this functionality in the Blazor Server and Blazor WAMS projects. Added an integration test to test it.
Updated the test verifier to scrub GUIDs to enable passed tests.
Updated the sample nav menu to point to the two new samples.